move things into ui and feat dirs
