--- src/lib/visualization/index.ts
@@ -1,2 +1,0 @@
-export { default as Visualization } from './Visualization';
-export * from './types';
